+++ docs/changelog.md
@@ -6,6 +6,7 @@
 * Added a network of consensus-layer clients
 * Hooked up genesis generation for Geth & CL nodes
 * Lighthouse nodes peer with each other
+* Add forkmon to the started network
 
 # 0.1.0
 * Initial commit